requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.9.2
python-multipart==0.0.12
pillow==10.4.0
httpx==0.27.2
google-genai==0.6.0
